Consolidate both types of deposits into the deposit module
