+++ lib/sdr_client.rb
@@ -9,7 +9,6 @@ require 'cocina/models'
 
 require 'sdr_client/version'
 require 'sdr_client/deposit'
-require 'sdr_client/model_deposit'
 require 'sdr_client/credentials'
 require 'sdr_client/login'
 require 'sdr_client/login_prompt'
